api/requirements.txt
numpy
Pillow
Flask